+++ src/api/mod.rs
@@ -1,1 +1,6 @@
 mod spotify_client;
+
+// tmp
+pub mod api;
+mod api_models;
+pub mod cache;
