+++ src/meson.build
@@ -29,11 +29,17 @@ run_command(
 )
 
 # find . -name "*.rs" -printf "'%p',\n"
-sources = files(
+sources = files('./api/api.rs',
+'./api/cache.rs',
+'./api/api.rs',
+'./api/api_models.rs',
+'./api/spotify_client.rs',
+'./api/mod.rs',
 './app/dispatch.rs',
 './app/components/details/details.rs',
 './app/components/details/details_model.rs',
 './app/components/details/mod.rs',
+'./app/components/artist/mod.rs',
 './app/components/utils.rs',
 './app/components/playback/playback_info.rs',
 './app/components/playback/playback_control.rs',
@@ -43,13 +49,13 @@ sources = files(
 './app/components/playlist_details/mod.rs',
 './app/components/album/album.rs',
 './app/components/album/mod.rs',
-'./app/components/artist/mod.rs',
 './app/components/artist_details/artist_details.rs',
 './app/components/artist_details/artist_details_model.rs',
 './app/components/artist_details/mod.rs',
 './app/components/login/login.rs',
 './app/components/login/mod.rs',
 './app/components/login/login_model.rs',
+'./app/components/window/mod.rs',
 './app/components/saved_playlists/saved_playlists_model.rs',
 './app/components/saved_playlists/mod.rs',
 './app/components/saved_playlists/saved_playlists.rs',
@@ -77,7 +83,6 @@ sources = files(
 './app/components/search/search_bar.rs',
 './app/components/search/mod.rs',
 './app/components/search/search.rs',
-'./app/components/window/mod.rs',
 './app/components/mod.rs',
 './app/loader.rs',
 './app/list_store.rs',
@@ -86,10 +91,7 @@ sources = files(
 './app/dbus/types.rs',
 './app/credentials.rs',
 './app/models.rs',
-'./app/backend/cache.rs',
-'./app/backend/api.rs',
 './app/backend/player.rs',
-'./app/backend/api_models.rs',
 './app/backend/mod.rs',
 './app/state/app_model.rs',
 './app/state/browser_state.rs',
@@ -98,8 +100,8 @@ sources = files(
 './app/state/app_state.rs',
 './app/state/mod.rs',
 './app/gtypes/album_model.rs',
-'./app/gtypes/artist_model.rs',
 './app/gtypes/song_model.rs',
+'./app/gtypes/artist_model.rs',
 './app/gtypes/mod.rs',
 './app/mod.rs',
 './config.rs',
